+++ src/Shapes2d/Triangle.h
@@ -7,7 +7,7 @@ namespace Shapes2d
 	{
 	public:
 		Triangle();
-		bool        IsPointWithinBounds(int x, int y);
+		bool        IsPointWithinBounds(float x, float y);
 		bool        DoesRectangleOverlap(int x1, int y1, int x2, int y2);
 		void        AffectVector(int x, int y, ofVec3f * actionVector, bool isRotation = false);
 		void        Draw();
